backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
django-filter==23.2
requests==2.31.0
flake8==6.0.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
pyTelegramBotAPI==4.14.0
